Build docker image, To #34515520
