--- cmd/webhook/app/add_scheduler.go
@@ -1,10 +1,0 @@
-package app
-
-import (
-	"github.com/fluid-cloudnative/fluid/pkg/webhook/scheduler/mutating"
-)
-
-func init() {
-	addHandlers(mutating.HandlerMap)
-	// addHandlers(validating.HandlerMap)
-}
